Homing Override
I mostly print off the right extruder, and was having failed print cancellations due to a slicer end g-code issue, when I finally got it to home it would home with the right extruder active and drive it into the bed. After destroying a couple beds I created a homing override which calls T0 macro prior to homing. Works well in my testing.